Fix metadata namespaces to be more intuitive
This is done in order to have a clear distinction between the
different layers.
